AI/requirements.txt
# Core Dependencies
beautifulsoup4==4.12.3
llama-cpp-python==0.2.32
numpy==1.26.3
requests==2.31.0
sentence-transformers==2.2.2
torch
flask>=2.2.5
werkzeug
tqdm==4.66.1
huggingface-hub==0.19.4
transformers==4.36.2
urllib3<2.0.0
gdown
pysmartdl>=1.3.4
icalendar==4.0.8

# NLP and Text Processing
nltk==3.8.1
python-dateutil==2.8.2
scikit-learn>=1.3.2

# Web Scraping and Content Extraction
newspaper3k==0.2.8
trafilatura==1.6.1
readability-lxml==0.8.1

# Knowledge Graph
networkx==3.2.1
